requirements.txt
pandas==2.1.4
numpy==1.26.4
openpyxl==3.1.2
sqlalchemy==2.0.25
xgboost==2.0.3
scikit-learn==1.4.0
mlflow==2.10.0
shap==0.44.1
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
joblib==1.3.2
anthropic==0.18.1
langchain==0.1.9
langchain-groq==0.1.3
langchain-community==0.0.24
